requirements.txt
langchain-core==1.0.3
langchain-google-genai==3.0.1
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
secure-smtplib==0.1.1
Markdown==3.10